changes db var name and fixes type for transaction

+++ src/controllers/transacoes/transacoes.go
@@ -9,7 +9,7 @@ import (
 )
 
 type TransactionPayload struct {
-	Value       int    `json:"valor"`
+	Value       float64    `json:"valor"`
 	Type        string `json:"tipo"`
 	Description string `json:"descricao"`
 }
